syncthing: Ignore permissions by default for new shared folders.
Utilize the newly introduced "folder default configuration" [1]
feature to preset the ignorePerms option to true.  This will not touch
any existing folders, but set the option when configuring a new folder
to be synced.

Specifically on DSM, the ignorePerms option should definitely be used
with every Syncthing folder.  Otherwise, it will possibly try to
override the UNIX permissions on synced files, which will reset the
ACL-based permissions on the content.  Possibly making them unusable
for the rest of the DSM services.

Note that this also only applies for new installations, because
config.xml is (and must be) kept during updates.

[1]: https://github.com/syncthing/syncthing/pull/7131